requirements.txt
altgraph==0.17
appdirs==1.4.4
astroid==2.4.2
backcall==0.2.0
click==7.1.2
colorama==0.4.4
dataclasses==0.6
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
importlib-metadata==3.1.1
install==1.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.9.1
mccabe==0.6.1
numpy==1.19.3
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
Pygments==2.7.2
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pylint==2.6.0
python-dotenv==0.15.0
pywin32-ctypes==0.2.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
